some cleanup and using the correct size to avoid issues

+++ saleaeinterface.h
@@ -41,8 +41,8 @@ unsigned int saleaeinterface_getchannelcount(saleaeinterface_t *obj);
 unsigned int saleaeinterface_getsamplerate(saleaeinterface_t *obj);
 void saleaeinterface_setsamplerate(saleaeinterface_t *obj, unsigned int rate);
 /* the user has to alloc the ptr variable and give it to the function */
-size_t saleaeinterface_getsupportedsamplerates(saleaeinterface_t *obj,
-                            unsigned int *ptr, size_t len);
+int saleaeinterface_getsupportedsamplerates(saleaeinterface_t *obj,
+                            unsigned int *ptr, unsigned int len);
 
 #ifdef __cplusplus
 } /* extern C end */
